Add files via upload
v0.1 - Originally intended to be the first upload, original files lost. The base website was was a basic html skeleton with snippets of css, mainly created as a course project for learning about webscraping later on. 

v0.2 - This/ first upload: This is the current state of the website, it includes a working header with nav bars and logo animation, with link icons leading to respective sites (profiles not setup yet) placeholder text has been added to ensure scrolling works as well as flex wrapping to ensure the header wraps neatly to smaller devices.